[Feat] Tooltip formatting (#1102)
* code review update
* move chicklet code to separate component
