+++ src/reducers/vis-state-merger.js
@@ -301,8 +301,8 @@ export function mergeInteractionTooltipConfig(state, tooltipConfig = {}) {
     } else {
       // if dataset is loaded
       const allFields = state.datasets[dataId].fields.map(d => d.name);
-      const foundFieldsToShow = tooltipConfig.fieldsToShow[dataId].filter(name =>
-        allFields.includes(name)
+      const foundFieldsToShow = tooltipConfig.fieldsToShow[dataId].filter(field =>
+        allFields.includes(field.name)
       );
 
       mergedTooltip[dataId] = foundFieldsToShow;
