+++ src/utils/interaction-utils.js
@@ -77,7 +77,7 @@ export const BRUSH_CONFIG = {
 export function findFieldsToShow({fields, id}) {
   // first find default tooltip fields for trips
   const fieldsToShow = DEFAULT_TOOLTIP_FIELDS.reduce((prev, curr) => {
-    if (fields.find(({name}) => curr === name)) {
+    if (fields.find(({name}) => curr.name === name)) {
       prev.push(curr);
     }
     return prev;
@@ -102,7 +102,12 @@ function autoFindTooltipFields(fields) {
       type !== 'object'
   );
 
-  return fieldsToShow.slice(0, MAX_DEFAULT_TOOLTIPS).map(d => d.name);
+  return fieldsToShow.slice(0, MAX_DEFAULT_TOOLTIPS).map(({name}) => {
+    return {
+      name,
+      format: null
+    };
+  });
 }
 
 function _mergeFieldPairs(pairs) {
